extract pool id to config to avoid circular dependencies
