+++ src/components/nodes/pool/config.js
@@ -1,0 +1,1 @@
+export const id = 'processmaker-modeler-pool';
