+++ src/components/shapeStackUtils.js
@@ -1,4 +1,4 @@
-import { id as poolId } from './nodes/pool';
+import { id as poolId } from './nodes/pool/config';
 import { id as laneId } from './nodes/poolLane/config';
 
 export default function ensureShapeIsNotCovered(shape, graph) {
